% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courtplot.R
\name{fullcourt}
\alias{fullcourt}
\title{Plotting an NBA Court}
\usage{
fullcourt()
}
\value{
returns ggplot object that can be used for NBA plots
}
\description{
Plotting an NBA Court
}
\examples{
fullcourt() # plots court that you can layer various geoms onto
}
